Add pipeline feedback improvements
